requirements.txt
# Core scientific library requirements
# These are also listed in pyproject.toml,
# but this file is useful for CI/CD and Docker.
numpy
biopython
MDAnalysis
scikit-learn
pyqubo
dwave-neal
dadapy
# idtxl Must install from source
pyyaml
redis
rq